+++ DawnLib/src/API/Terminal/ITerminalPurchase.cs
@@ -1,4 +1,6 @@
-﻿namespace Dawn;
+﻿using System.Linq;
+
+namespace Dawn;
 public interface ITerminalPurchase
 {
     IProvider<int> Cost { get; }
@@ -20,5 +22,33 @@ internal class ConstantTerminalPredicate(TerminalPurchaseResult result) : ITermi
     public TerminalPurchaseResult CanPurchase()
     {
         return result;
+    }
+}
+
+internal class LethalLevelLoaderTerminalPredicate(object extendedLevelObject) : ITerminalPurchasePredicate
+{
+    private TerminalNode? _failNode = null;
+    public TerminalPurchaseResult CanPurchase()
+    {
+        LethalLevelLoader.ExtendedLevel extendedLevel = (LethalLevelLoader.ExtendedLevel)extendedLevelObject;
+        if (extendedLevel.IsRouteLocked && extendedLevel.IsRouteHidden)
+        {
+            return TerminalPurchaseResult.Hidden().SetFailure(true);
+        }
+        else if (extendedLevel.IsRouteHidden)
+        {
+            return TerminalPurchaseResult.Hidden().SetFailure(false);
+        }
+        else if (extendedLevel.IsRouteLocked)
+        {
+            if (_failNode == null)
+            {
+                _failNode = new TerminalNodeBuilder($"{extendedLevel.name.Replace(" ", "").SkipWhile(x => !char.IsLetter(x)).ToArray()}LethalLevelLoaderTerminalPredicateFail")
+                    .SetDisplayText($"{extendedLevel.LockedRouteNodeText}")
+                    .Build();
+            }
+            return TerminalPurchaseResult.Fail(_failNode);
+        }
+        return TerminalPurchaseResult.Success();
     }
 }